Added comments and cleaned up App.tsx
