+++ src/components/example.tsx
@@ -1,0 +1,67 @@
+/**
+ * @file Contains the App top level component.
+ */
+import React, { useState } from 'react';
+import Button from '@mui/material/Button';
+import reactLogo from '../assets/react.svg';
+
+import '../App.css';
+import { useWasm } from '../hooks/wasm';
+import CalculatorModule from '../cpp/Calculator';
+
+/**
+ * Top level component.
+ *
+ * @returns top level component
+ */
+function Example(): JSX.Element {
+  const [count, setCount] = useState(0);
+  const calcModule = useWasm(CalculatorModule);
+
+  return (
+    <div className="Example">
+      <div>
+        <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
+          <img src="/vite.svg" className="logo" alt="Vite logo" />
+        </a>
+        <a href="https://reactjs.org" target="_blank" rel="noreferrer">
+          <img src={reactLogo} className="logo react" alt="React logo" />
+        </a>
+      </div>
+      <h1>Vite + React</h1>
+      <div className="card">
+        <button
+          type="button"
+          onClick={() => {
+            setCount(() => {
+              if (count >= 10) {
+                return new calcModule.Calculator().subtract(count, count);
+              }
+              return calcModule.increment(count);
+            });
+          }}
+        >
+          count is {count}
+        </button>
+        <Button
+          color="secondary"
+          type="button"
+          onClick={() => setCount((currentCount) => currentCount + 1)}
+        >
+          count is {count}
+        </Button>
+        <Button color="primary" type="button">
+          Just a visual MUI button
+        </Button>
+        <p>
+          Edit <code>src/App.tsx</code> and save to test HMR
+        </p>
+      </div>
+      <p className="read-the-docs">
+        Click on the Vite and React logos to learn more
+      </p>
+    </div>
+  );
+}
+
+export default Example;
