+++ src/main.tsx
@@ -5,10 +5,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import PersistentDrawerLeft from './components/menu'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <PersistentDrawerLeft />
+    <App />
   </React.StrictMode>
 );
